+++ webots_ros2_driver/webots_ros2_driver/webots_launcher.py
@@ -17,11 +17,20 @@
 """This launcher simply starts Webots."""
 
 import os
+import re
+import shutil
 import sys
+import tempfile
+
 from launch.actions import ExecuteProcess
+from launch_ros.actions import Node
+from launch.launch_context import LaunchContext
 from launch.substitution import Substitution
 from launch.substitutions import TextSubstitution
-from webots_ros2_driver.utils import get_webots_home, handle_webots_installation
+from launch.substitutions.path_join_substitution import PathJoinSubstitution
+
+from webots_ros2_driver.utils import (get_webots_home,
+                                      handle_webots_installation)
 
 
 class _ConditionalSubstitution(Substitution):
@@ -48,7 +57,11 @@ class WebotsLauncher(ExecuteProcess):
         webots_path = os.path.join(webots_path, 'webots')
 
         mode = mode if isinstance(mode, Substitution) else TextSubstitution(text=mode)
-        world = world if isinstance(world, Substitution) else TextSubstitution(text=world)
+
+        self.__world_copy = tempfile.NamedTemporaryFile(mode='w+', suffix='_world_with_URDF_robot.wbt', delete=False)
+        self.__world = world
+        if not isinstance(world, Substitution):
+            world = TextSubstitution(text=self.__world_copy.name)
 
         no_rendering = _ConditionalSubstitution(condition=gui, false_value='--no-rendering')
         stdout = _ConditionalSubstitution(condition=gui, false_value='--stdout')
@@ -56,6 +69,7 @@ class WebotsLauncher(ExecuteProcess):
         minimize = _ConditionalSubstitution(condition=gui, false_value='--minimize')
         stream_argument = _ConditionalSubstitution(condition=stream, true_value='--stream')
         xvfb_run_prefix = []
+
         if 'WEBOTS_OFFSCREEN' in os.environ:
             xvfb_run_prefix.append('xvfb-run')
             xvfb_run_prefix.append('--auto-servernum')
@@ -75,5 +89,71 @@ class WebotsLauncher(ExecuteProcess):
                 '--batch',
                 ['--mode=', mode],
             ],
+            **kwargs
+        )
+
+    def execute(self, context: LaunchContext):
+        # User can give a PathJoinSubstitution world or an absolute path world
+        if isinstance(self.__world, PathJoinSubstitution):
+            world_path = self.__world.perform(context)
+            context.launch_configurations['world'] = self.__world_copy.name
+        else:
+            world_path = self.__world
+
+        shutil.copy2(world_path, self.__world_copy.name)
+
+        # Update relative paths in the world
+        with open(self.__world_copy.name, 'r') as file:
+            content = file.read()
+
+        for match in re.finditer('url\s*\[?\s*\"(.*?)\"', content):
+            url_path = match.group(1)
+
+            # Absolute path or Webots relative path or Web paths
+            if os.path.isabs(url_path) or 'webots://' in url_path or 'http://' in url_path or 'https://' in url_path:
+                continue
+
+            new_url_path = '"' + os.path.split(world_path)[0] + '/' + url_path + '"'
+            url_path = '"' + url_path + '"'
+            content = content.replace(url_path, new_url_path)
+
+        with open(self.__world_copy.name, 'w') as file:
+            file.write(content)
+
+        # Add the Ros2Supervisor
+        indent = '  '
+        world_file = open(self.__world_copy.name, 'a')
+        world_file.write('Robot {\n')
+        world_file.write(indent + 'name "Ros2Supervisor"\n')
+        world_file.write(indent + 'controller "<extern>"\n')
+        world_file.write(indent + 'supervisor TRUE\n')
+        world_file.write('}\n')
+        world_file.close()
+
+        return super().execute(context)
+
+    def _shutdown_process(self, context, *, send_sigint):
+        # Remove copy of the world and the corresponding ".wbproj" file
+        if self.__world_copy:
+            self.__world_copy.close()
+            if os.path.isfile(self.__world_copy.name):
+                os.unlink(self.__world_copy.name)
+
+            path, file = os.path.split(self.__world_copy.name)
+            world_copy_secondary_file = os.path.join(path, '.' + file[:-1] + 'proj')
+            if os.path.isfile(world_copy_secondary_file):
+                os.unlink(world_copy_secondary_file)
+        return super()._shutdown_process(context, send_sigint=send_sigint)
+
+
+class Ros2SupervisorLauncher(Node):
+    def __init__(self, output='screen', respawn=True, **kwargs):
+        # Launch the Ros2Supervisor node
+        super().__init__(
+            package='webots_ros2_driver',
+            executable='ros2_supervisor.py',
+            output=output,
+            additional_env={'WEBOTS_ROBOT_NAME': 'Ros2Supervisor'},
+            respawn=respawn,
             **kwargs
         )