+++ webots_ros2_tests/test/test_system_turtlebot_tutorial_navigation.py
@@ -72,7 +72,7 @@ def generate_test_description():
     ])
 
 
-class TestTurtlebotTutorials(TestWebots):
+class TestTurtlebotNavigationTutorial(TestWebots):
     @classmethod
     def setUpClass(cls):
         rclpy.init()